Move bulk data processing functions to be methods
Starting to grok the borrow checker. Maybe. Making structs from the deserialized ones,
added this level of indirection because I couldn't figure out how to make Serde not
freakout with the http requests on the structs. It would be nice to figue out how to
make structs that wrap the other structs more gracefully, something like derrived data,
or *gasp* struct inheritance? genericslol?

It took a while to get the mutability and borrow annotations correct.
(Part of) The solution wound up being to make the top level instanciation let mut,
apparently in debugging these issues start futher up the chain?